+++ CHANGES.md
@@ -1,5 +1,6 @@
 - Disable cache usage in Restic check profiles
 - Show stale tasks in verbose mode only
+- Pick the innermost Bitferry volume in case of volumes nesting
 - Fix an options profile application bug
 
 ## 0.0.4